requirements.dev.txt
Tox
PyTest
WeMake-Python-StyleGuide
ISort<5
Coverage
Sphinx
psycopg2-binary
PySpark-Stubs
